rewrite: move calculation of set to rebase to `MutableRepo`
This lets us make `parent_mapping` private again.
